// frontend/lib/api/generated/model/racionAnualOut.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SIGA-UNMSM API
 * API del Sistema Integral de Gestión de Almacén y Abastecimiento Alimentario — UNMSM. Módulos activos: Autenticación, Usuarios, Almacenes, Catálogo Nutricional.
 * OpenAPI spec version: 0.1.0
 */

export interface RacionAnualOut {
  racion_anual_id: number;
  sede_id: number;
  centro_consumo_id: number;
  anio: number;
  poblacion_atendida: number;
  raciones_dia: number;
  estado: string;
  creado_en: string;
}